Change quotes in locale.

+++ chip8/memory.v
@@ -23,7 +23,7 @@ pub fn (mut self Memory) save_byte(address u16, value u8) {
 // Copies an array of bytes directly to the memory at the specified offset.
 pub fn (mut self Memory) copy_bytes(offset u16, bytes []u8) {
 	for index := 0; index < bytes.len; index++ {
-		self.save_byte(offset + index, bytes[index])
+		self.save_byte(u16(offset + index), bytes[index])
 	}
 }
